refactor * logic for formatting to markdown in separate class
