+++ dit-calendar-bot/src/main/kotlin/com/ditcalendar/bot/DitCalendarCommand.kt
@@ -7,6 +7,7 @@ import com.ditcalendar.bot.endpoint.TaskEndpoint
 import com.ditcalendar.bot.error.DitBotError
 import com.ditcalendar.bot.error.InvalidRequest
 import com.ditcalendar.bot.error.UnassigmentError
+import com.ditcalendar.bot.markdown.toStringInMarkdown
 import com.elbekD.bot.types.Message
 import com.elbekD.bot.types.User
 import com.github.kittinunf.fuel.core.FuelError
