+++ dit-calendar-bot/src/main/kotlin/com/ditcalendar/bot/data/core/Base.kt
@@ -7,6 +7,4 @@ abstract class Base {
     var entryId: Long = 0
 
     var version: Int = 0
-
-    abstract fun toStringInMarkdown(): String
 }
